updated all NetworkService callbacks
