+++ QuizApp.xcworkspace/xcuserdata/matejbutkovic.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "9D666A0C-C34B-4F79-90B9-72C2508FCAE3"
+   type = "0"
+   version = "2.0">
+</Bucket>
