+++ QuizApp/LoginView/LoginViewController.swift
@@ -47,18 +47,13 @@ class LoginViewController: UIViewController {
         loginButton.isEnabled = false
         let email = emailField.text ?? ""
         let password = passwordField.text ?? ""
-//        let response = dataService.login(email: email, password: password)
         presenter.login(email: email, password: password)
-        
-
     }
 
     private func addConstraints() {
 
         let elements = [titleLabel, emailField, passwordField, errorLable, loginButton]
-
         let height = view.bounds.height
-
         for element in elements {
             element!.snp.makeConstraints { make in
                 // should gard this instead of force unwrapping
@@ -67,9 +62,7 @@ class LoginViewController: UIViewController {
                 make.centerX.equalToSuperview()
             }
         }
-//        titleLabel.snp.makeConstraints { make in
-//            make.top.equalTo(view).offset(height * 0.1)
-//        }
+
         titleLabel.isHidden = true
 
         emailField.snp.makeConstraints { make in
@@ -80,7 +73,6 @@ class LoginViewController: UIViewController {
             make.top.equalTo(emailField).offset(45 + 18)
         }
 
-
         errorLable.snp.makeConstraints { make in
             make.top.equalTo(passwordField).offset(38)
         }
@@ -89,7 +81,6 @@ class LoginViewController: UIViewController {
             make.top.equalTo(errorLable).offset(38)
         }
         loginButton.layer.cornerRadius = 20
-
 
     }
 
@@ -102,7 +93,7 @@ extension LoginViewController: LoginPresenterDelegate {
         errorLable.isHidden = true
         presenter.showQuizzes()
     }
-    
+
     func loginFail() {
         loginButton.isEnabled = true
         print("Error while login")
